requirements.txt
Django>=4.2
psycopg2-binary
celery
redis
django-celery-beat
django-celery-results
python-dotenv
djangorestframework
requests